requirements.txt
amqp==5.2.0
astroid==3.1.0
billiard==4.2.0
black==24.4.2
blinker==1.8.1
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dill==0.3.8
dnspython==2.6.1
exceptiongroup==1.2.1
Flask==3.0.3
gunicorn==22.0.0
importlib_metadata==7.1.0
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.3
kombu==5.3.7
MarkupSafe==2.1.5
mccabe==0.7.0
mongoengine==0.28.2
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.1
pluggy==1.5.0
prompt_toolkit==3.0.47
pylint==3.1.0
pymongo==4.7.1
pytest==8.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
tomli==2.0.1
tomlkit==0.12.4
typing_extensions==4.11.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.2
zipp==3.18.1
